+++ MaterialDesignExtensions/Model/IAutocompleteSource.cs
@@ -16,4 +16,16 @@ namespace MaterialDesignExtensions.Model
     {
         new IEnumerable<T> Search(string searchTerm);
     }
+
+    public abstract class AutocompleteSource<T> : IAutocompleteSource<T>
+    {
+        public AutocompleteSource() { }
+
+        public abstract IEnumerable<T> Search(string searchTerm);
+
+        IEnumerable IAutocompleteSource.Search(string searchTerm)
+        {
+            return Search(searchTerm);
+        }
+    }
 }