+++ ReleaseNotes.md
@@ -83,5 +83,7 @@
 #### Features
 * Improved layout for file system controls
 * New control `TabControlStepper`
+* Improved layout for `OversizedNumberSpinner`
+* New interface `IAutocompleteSourceChangingItems` to propagate source changes to `Autocomplete`
 #### Fixes
 * Handle `PathTooLongException`